修复页码跳转问题

+++ book_vue/src/views/BookManage.vue
@@ -36,12 +36,30 @@ import axios from 'axios'
 export default {
   methods: {
     deleteBook(row) {
-      axios.delete("http://localhost:9999/book/delete/" + row.id).then(ret => {
-        this.$alert("《" + row.name + "》已删除", {
-          callback: action => {
-            this.$router.go(0);
+      this.$confirm('确定要删除该图书吗?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        axios.delete("http://localhost:9999/book/delete/" + row.id).then(ret => {
+          this.$message.success("《" + row.name + "》已删除");
+          
+          // 计算删除后的总页数
+          const newTotal = this.total - 1;
+          const newTotalPages = Math.ceil(newTotal / this.pageSize);
+          
+          // 如果当前页大于新的总页数，就跳转到最后一页
+          if (this.currentPage > newTotalPages) {
+            this.currentPage = Math.max(1, newTotalPages);
           }
+          
+          // 重新获取当前页数据
+          this.fetchData(this.currentPage);
+        }).catch(() => {
+          this.$message.error('删除失败');
         });
+      }).catch(() => {
+        // 取消删除操作
       });
     },
     update(row) {
